llama_lab/requirements.txt
llama-cpp-python==0.3.10
streamlit
